% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extensions.R
\name{get_extensions}
\alias{get_extensions}
\title{Fetch and Parse Quarto Extension Listing YAML Files from GitHub}
\usage{
get_extensions()
}
\value{
A dataframe that represents the parsed content of the extension YAML files
and includes the file name as an additional column.
}
\description{
Fetch and Parse Quarto Extension Listing YAML Files from GitHub
}
\examples{
\dontrun{
 get_extensions()
}
}
