% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_url.R
\name{github_url_parts}
\alias{github_url_parts}
\title{Get GitHub URL Components}
\usage{
github_url_parts(urls)
}
\arguments{
\item{urls}{A string vector with GitHub URLs}
}
\value{
A dataframe containing the components of a GitHub URL
}
\description{
Get GitHub URL Components
}
\examples{
 quartoextexp:::github_url_parts("http://github.com/tidyverse/dplyr")
 quartoextexp:::github_url_parts("https://github.com/tidyverse/dplyr.git")
}
