% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_url.R
\name{github_url_to_repo}
\alias{github_url_to_repo}
\title{Extract User and Repository from GitHub URL}
\usage{
github_url_to_repo(url)
}
\arguments{
\item{url}{The GitHub URL to extract the user and repository name from}
}
\value{
A character vector containing the user and repository name
}
\description{
Extract User and Repository from GitHub URL
}
\examples{
quartoextexp:::github_url_to_repo("http://github.com/tidyverse/dplyr")
quartoextexp:::github_url_to_repo("https://github.com/tidyverse/dplyr.git")
}
