% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_url.R
\name{github_url_to_user}
\alias{github_url_to_user}
\title{Extract Username from GitHub URL}
\usage{
github_url_to_user(url)
}
\arguments{
\item{url}{The GitHub URL to extract the user name from}
}
\value{
A character vector containing the user name
}
\description{
Extract Username from GitHub URL
}
\examples{
quartoextexp:::github_url_to_user("http://github.com/tidyverse/dplyr")
quartoextexp:::github_url_to_user("https://github.com/tidyverse/dplyr.git")
}
