% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{githubURLParts}
\alias{githubURLParts}
\title{Get GitHub URL Components}
\usage{
githubURLParts(urls)
}
\arguments{
\item{urls}{A string vector with GitHub URLs}
}
\value{
A dataframe containing the components of a GitHub URL
}
\description{
Get GitHub URL Components
}
\examples{
 quartoextexp:::githubURLParts("http://github.com/tidyverse/dplyr")
 quartoextexp:::githubURLParts("https://github.com/tidyverse/dplyr.git")
}
