% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run Quarto Extension Explorer Shiny App}
\usage{
run_app(install_dir = getwd())
}
\arguments{
\item{install_dir}{The installation directory for the Quarto extension
by default this is the current working directory}
}
\description{
Run Quarto Extension Explorer Shiny App
}
\examples{
\dontrun{
 run_app()
}
}
